backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
